+++ src/OpenApiRuntime/Client/Psr7HttplugEndpoint.php
@@ -4,15 +4,6 @@ declare(strict_types=1);
 
 namespace Jane\OpenApiRuntime\Client;
 
-use Psr\Http\Message\ResponseInterface;
-use Symfony\Component\Serializer\SerializerInterface;
-
-interface Psr7HttplugEndpoint extends Endpoint
+interface Psr7HttplugEndpoint extends Psr7Endpoint
 {
-    /**
-     * Parse and transform a PSR7 Response into a different object.
-     *
-     * Implementations may vary depending the status code of the response and the fetch mode used.
-     */
-    public function parsePSR7Response(ResponseInterface $response, SerializerInterface $serializer, string $fetchMode = Client::FETCH_OBJECT);
 }